rust175: add a rust package for rust 1.75.0.
This has so far been verified to build natively on
NetBSD/amd64 10.99.4, and also that it can cross-build for
the other NetBSD targets.

The source for the NetBSD/amd64 bootstrap kit is chnged
to come from NetBSD, not from rust, due to problems with
rust's bootstrap kit (stage1 not finding libLLVM-16.so,
so it's built with an external LLVM and not with semi-static
linking).  Issue forthcoming (still sitting in a browser on my
PC at work, it seems...)