qrcp: transfers files over Wi-Fi with a QR code
qrcp transfers files over Wi-Fi from your computer to a mobile
device by scanning a QR code without leaving the terminal.  qrcp
binds a web server to the address of your Wi-Fi network interface
on a random port and creates a handler for it. The default handler
serves the content and exits the program when the transfer is
complete. When used to receive files, qrcp serves an upload page
and handles the transfer.