lang/gcc13-gnat: add GCC 13.2 GNAT compiler & tools
The GNU Compiler Collection (GCC). This package contains GNAT compiler and Ada tools

This package is based on gcc13 patches for C/C++/Fortran compilers and gcc10-aux patches for Ada compiler.

The package is tightly bound to wip/gcc13-gnat subdirectory (in buildlink3.mk and gcc.mk files)
and this path must be replaced with lang/gcc13-gnat when/if moving the package to pkgsrc.
gcc.mk file added as a replacment for pkgsrc's compiler/mk/gcc.mk file, which still doesn't know about gcc10-aux,
which in turn needed for building. Although Jonathan Perkin points out that GCC_REQD is not related to the AUX
versions of GCC, pkgsrc's gcc.mk contains a special treatment of gcc6-aux.
It is also assumed that packages that will use this package will use provided gcc.mk .