cyclone-bootstrap: adjust installation of files for bootstrapping
The new file locations make it possible to install the bootstrap
version and cyclone at the same time.

These changes are in preparation for importing the full cyclone
package, and this package will be used for the bootstrapping stage
only.  By keeping the bootstrapping package and main package
separate, the patches for cyclone will be simpler, since the
generated C files will not need to be patched.