prusaslicer: update to release 2.7.2-rc2.
This is the second release candidate of PrusaSlicer 2.7.2, which fixes bugs found in the previous release candidate. Please, read the change logs of 2.7.2-alpha2, 2.7.2-beta1 and 2.7.2-rc1 for the complete list of bugfixes and improvements over 2.7.1.

The release candidate saves its profiles into regular PrusaSlicer configuration directory. When you first run it, it will search for all configurations produced by alpha or beta versions and offer to create a copy of the latest one.
Bug fixes with respect to 2.7.2-rc1

-   When using the "No sparse layers" wipe tower, the print head sometimes lowered before travelling to the wipe tower, digging into an object and destroying the print (#12330).
-   Fixed a crash when loading certain AMF files (#12332).

Localization

-   Updated JA dictionary.