prusaslicer: update to release 2.7.2
This is the stable release of PrusaSlicer 2.7.2, fixing one bug found in the previous release candidate. Please, read the change logs of 2.7.2-alpha2, 2.7.2-beta1, 2.7.2-rc1 and 2.7.2-rc2 for the complete list of bugfixes and improvements over 2.7.1.

If any of the PrusaSlicer 2.7.2 alphas or betas was used before on the same machine, PrusaSlicer 2.7.2 will offer to import such alpha or beta configuration when it is first executed.

Bugs fixed with respect to 2.7.1

-   When specific extruder was set for the wipe tower and z position has changed in custom toolchange G-code, the z position was not correctly accounted for.