Import Mobius 0.12.2 as go-mobius
Mobius is a cross-platform command line Hotline client and server implemented
in Golang. The goal of Mobius server is to make it simple to run a Hotline
server on macOS, Linux, and Windows, with full compatibility for all popular
Hotline clients.

(I haven't been able to build it fully just yet)