prusaslicer: Update to 2.7.5-rc1
PrusaSlicer 2.7.5-rc1 Pre-release

Summary

This is the first release candidate of PrusaSlicer 2.7.5. This release adds
support for new features introduced with firmware version 1.8.0 for Prusa SLA
printers.

Note that this release candidate uses the same configuration folder as the
regular releases.

Extended SLA layer separation configuration

In previous versions of SL1/SL1S firmware, the layer separation configuration
mostly consisted of selecting a 'Material printing profile' from three
predefined options (Slow / Fast / High viscosity). Based on the selection, the
printer would use an internally defined set of parameters to perform the layer
separation. To make it possible to fine-tune the layer separation procedure for
each material, new 1.8.0-beta.0 firmware for the SL1/SL1S printers allows more
detailed control over the parameters of the separation. You can read more about
this change in Prusa SLA FW 1.8.0-beta.0 change log.

To keep up with these changes, PrusaSlicer 2.7.5 changes the way how the
separation is configured: the 'Material printing profile' selector is no longer
shown and many new parameters are now accessible in
Material Settings -> Material printing profile. Note that this change is only
applied for Prusa SL1 and SL1S. With all the other printers, everything stays
as it was in 2.7.4.

Old projects using the 'Material printing profile' selection are automatically
converted when loaded, and the values for all the parameters are set to match
the legacy profile selected in the project. On the other hand, opening a 3MF
project from 2.7.5 or higher in an older version of PrusaSlicer will silently
ignore the new parameters and the value of 'Material printing profile' (which
is retained internally and kept in our system profiles) will be used instead.

Other improvements and fixes with respect to 2.7.4

- An extra legend window was added into the SLA preview. It shows when using
  the vertical slider and contains information about current layer area and
  time estimate.
- Parameter 'Printer absolute correction' was added into Material Overrides.
- Parameter 'Printer scaling correction' was removed from Material Overrides.
- Fixed UI glitches in SLA Material Overrides tab when the window was moved
  from one monitor to another with different scaling factor, or when
  dark/light mode was switched.

Translations

- Updated CS, PL, DE, IT, ES, FR, JA dictionaries.