u-boot-tools: prefix tool names with u-boot-.
The u-boot tools include a mkimage program, which has the same name as
a native NetBSD program (although the latter is not installed).  This
creates at least the possibility of a conflict and in any case is not
a very specific, informative name.  For the purposes of this package,
change the names to be more informative by adding a u-boot- prefix.
This seems to have been the consensus of the discussion on tech-pkg:
https://mail-index.netbsd.org/tech-pkg/2023/08/28/msg028077.html.