prusaslicer: update from 2.8.0beta1 to 2.8.0rc2.
PrusaSlicer 2.8.0-rc2

This is the second release candidate of PrusaSlicer 2.8.0. This release
candidate fixes several bugs found in previous 2.8.0-rc1.

The release candidate saves its profiles into regular PrusaSlicer configuration
directory. When you first run it, it will search for all configurations
produced by alpha or beta versions and offer to create a copy of the latest
one.

Improvements with respect to 2.8.0-rc1

- The page in ConfigWizard with profiles sources selection now contains
  descriptive text and links to the website where configuration updates can be
  downloaded.

Bug fixes with respect to 2.8.0-rc1

- Thumbnails in Connect view did not display.
- Fixed incorrect colors in Preview when adding Color Changes to a painted
  object.

Translations

- Updated phrases and POT
- Updated internal dictionaries (CS, PL, DE, IT, ES, FR, JA)

Architecture, infrastructure

AppImage for Ubuntu 24.04

An AppImage running on Ubuntu 24.04 is finally provided (unlike previous
releases). The problem was that PrusaSlicer now depends on WebKit library,
which is a big library with many dependencies. The new Ubuntu 24.04 uses new
version of WebKit which is incompatible with our binaries. Bundling WebKit into
the AppImage has proved to be difficult, because it has many dependencies and
bundling them too leads to many other issues.

Therefore, instead of bundling all the libraries into a single huge AppImage,
we provide two separate AppImages. One depends on the old WebKit (which is
currently shipped with almost all relevant Linux distributions), the other one
is built on Ubuntu 24.04 and links to the new WebKit.

To keep the confusion to a minumum, we do not provide the tarballs anymore,
just the two AppImages. They are able to extract themselves when run with
--appimage-extract command line parameter.

PrusaSlicer 2.8.0-rc1

This is the first release candidate of PrusaSlicer 2.8.0. This release
candidate fixes bugs found in previous 2.8.0-beta1.

The release candidate saves its profiles into regular PrusaSlicer configuration
directory. When you first run it, it will search for all configurations
produced by alpha or beta versions and offer to create a copy of the latest one.

An important note for Linux users: The provided binaries will not run on newest
Ubuntu 24.04 (and possibly other latest distributions). Solving problems with
all dependencies is more work than originally expected. We hope to get it
sorted for the 2.8.0 final release.

Improvements with respect to 2.7.5-rc1

- SLA specific: Added correction for cross layer curing. When printing using the
  MSLA technology, the UV light cures not only the layer it is "supposed" to,
  but it also reaches couple of already printed layers. This leads to
  deformations of the print, especially at overhangs. There is a now new
  parameter in Material Settings. It states how many layers are be affected by
  the cross layer curing for the particular Material, and PrusaSlicer is able to
  compensate for this effect by "deforming" the slices in the opposite way.

Bug fixes with respect to 2.8.0-beta1

- macOS specific: Fixed performance issues when run as standalone G-code Viewer.
- Toolbar size set in Preferences was not taken into account (#12838 (comment)).
- Fixed crash when selecting Color Print view in Preview (#12881).
- Fixed behavior of vertical slider: Color Change was added after moving slider
  when it was moved to its original position (#12877).
- macOS specific: Some keyboard shortcut did not work (#12874).
- Linux specific: The application stays logged in to Prusa Account when it is
  restarted.
- Loading of SLA archives did not work when path to the archive contained
  non-ASCII characters (this was actually already fixed in 2.8.0-alpha5).

Translations

- Updated POT
- Updated internal dictionaries (CS, PL, DE, IT, ES, FR, JA)
- Added Slovenian translation (#12817, thanks to @jernejp21)