py-dask: Update to 2024.6.2
2024.6.2

This is a patch release to update an issue with dask and distributed version pinning in the 2024.6.1 release.
    Get docs build passing (GH#11184) James Bourbeau
    profile._f_lineno: handle next_line being None in Python 3.13 (GH#8710) Adam Williamson

2024.6.1
Highlights

This release includes a critical fix that fixes a deadlock that can arise when dependencies of root-ish tasks are rescheduled, e.g. due to a worker being lost.
See GH#8703 by Hendrik Makait for more details.
    Cache global query-planning config (GH#11183) Richard (Rick) Zamora
    Python 3.13 fixes (GH#11185) Adam Williamson
    Fix test_map_freq_to_period_start for pandas=3 (GH#11181) James Bourbeau
    Bump release-drafter/release-drafter from 5 to 6 (GH#8699)