rust178: turn off thread local storage for NetBSD/aarch64*.
NetBSD aarch64 has a bug in the thread-local storage implementation,
ref. http://gnats.netbsd.org/58154, and this is responsible for the
build misery experienced earlier, ref.
https://github.com/rust-lang/rust/issues/123551
Therefore: turn it off for now on arm64.

Ideally one could check whether the specific OS version has the fix
or not, but e.g. __NetBSD_Version__ isn't easily available here
that I know, and this goes against the pattern that OS version
should not matter as long as it's >= supported version.
So until the fix for NetBSD PR#58154 has propagated to NetBSD
releases, and the old ones are no longer supported for pkgsrc
(which is going to take a while), this is what we do.