go123: final version, 1.23.0!
The latest Go release, version 1.23, arrives six months after Go 1.22. Most of
its changes are in the implementation of the toolchain, runtime, and libraries.
As always, the release maintains the Go 1 promise of compatibility. We expect
almost all Go programs to continue to compile and run as before.

Go 1.23 makes the (Go 1.22) “range-over-func” experiment a part of the
language.

Starting in Go 1.23, the Go toolchain can collect usage and breakage statistics
that help the Go team understand how the Go toolchain is used and how well it
is working. We refer to these statistics as Go telemetry.

Go telemetry is an opt-in system, controlled by the go telemetry command. By
default, the toolchain programs collect statistics in counter files that can be
inspected locally but are otherwise unused (go telemetry local).

Setting the GOROOT_FINAL environment variable no longer has an effect.
Distributions that install the go command to a location other than
$GOROOT/bin/go should install a symlink instead of relocating or copying the go
binary.