trurl: Import trurl-0.15 as wip/trurl
trurl is a tool in a similar spirit of tr but for URLs. Here, 'tr'
stands for translate or transpose.

trurl is a command line tool that parses and manipulates URLs, designed
to help shell script authors everywhere.

URLs are tricky to parse and there are numerous security problems in
software because of this. trurl wants to help soften this problem by
taking away the need for script and command line authors everywhere to
re-invent the wheel over and over.

trurl uses libcurl URL parser and will thus parse and understand URLs
exactly the same as curl the command line tool does - making it the
perfect companion tool.