py-snmp: Update to 6.2.5
6.2.5 has significant changes from the now-deprecated 4.  Because this
is wip, I'm not worrying about it, and assuming no one is using this -
or they would have updated.  6.2.5 removes asyncore; now ascyncio is
the true path.

Further details at: https://docs.lextudio.com/pysnmp/upgrade

(Update is to 6.2.5 rather than 7, because Home Assistant (this week)
is using 6.2.5, and this update is to help me debug HA issues.)