Add a package for cligen version 7.1.0.
CLIgen is a Command-Line Interface generator.

Well, actually it is not really a generator, since it does not
generate code for CLI:s. Instead, it builds and interprets
datastructures (a parse-tree) which a library (libcligen) interprets
in runtime. It is fast and efficient and helps you develop CLI:s
easier. You enter a CLI syntax in a text file, and write callback
functions in C. The callback functions add the semantics, that is,
what the commands in the CLI are supposed to do.