lwtools: update to version 4.23
2024-08-03: LWTOOLS 4.23. This version fixes a stack corruption bug related to
the object file output format. If you have been having random unexplained
crashes from lwasm, you definitely want this update. Also comes with the shiny
new ability to use a 0b or 0B prefix for binary constants.

2023-12-22: LWTOOLS 4.22. This version comes with a couple of useful additions
including being able to specify STDIN as the source file using - and the
ability to specify an offset and length for includebin. Of some use might be
the --no-warn=ifp1 flag to suppress warnings about the use of ifp1. Finally,
the default compile optimization flag is reduced to -O2. It seems there are
still compilers that actually generate incorrect code under -O3.

2023-04-23: LWTOOLS 4.21. This version comes with a couple of bugfixes related
to building with certain Windows tools as well as an undefined memory access
problem. Updating is recommended.