ti99sim: SDL-based TI-99/4a simulator
An emulator for the Texas Instruments TI-99/4A Home Computer. It
features full sound speech emulation. Supports emulation of the Gram
Kracker(TM), TI-Disk system, and 32K Memory expansion unit. This is a
port of a Win32 based emulator that I wrote a while back and am in the
process of porting to Linux using SDL (Simple DirectMedia Layer). The
code is now machine independant and should run on any platform
supported by SDL.