terraform: Make providers usable by OpenTofu
Terraform Registry (registry.terraform.io) could not be used by OpenTofu
because that would violate the Terraform Registry Terms of Use.

While in pkgsrc we do not use Terraform Registry but all the
terraform-provider-* are intended to be used as `filesystem_mirror`
OpenTofu could not get them in any ways without patching (but by
patching it we can violate the Terraform Registry Terms of Use).

Create a symbolic link to registry.opentofu.org so that they can be
used by OpenTofu too.

Now terraform-provider-* are usable by both Terraform and OpenTofu.