Import py-huggingface-hub-0.26.1 as wip/py-huggingface-hub
The huggingface_hub library allows you to interact with the Hugging
Face Hub, a machine learning platform for creators and collaborators.
Discover pre-trained models and datasets for your projects or play with
the hundreds of machine learning apps hosted on the Hub. You can also
create and share your own models and datasets with the community. The
huggingface_hub library provides a simple way to do all these things
with Python.