uv: Add uv version 0.5.8
TODO: Fails to build because the crate has two git dependencies.

uv is an extremely fast Python package and project manager written in
Rust.  It aims to replace a variety of tools such as pip, pip-tools,
pipx, poetry, pyenv, twine, and virtualenv with a single, unified
interface.

uv features 10-100x faster performance compared to pip, and provides
installation and management of Python versions, running and installing
Python applications, and support for inline dependency metadata in
scripts.  It offers comprehensive project management with a universal
lockfile, a pip-compatible interface for a performance boost with a
familiar CLI, and support for Cargo-style workspaces for scalable
projects. uv is also disk-space efficient, with a global cache for
dependency deduplication.