flux2: Update to 2.4.0
Changes:
2.4.0
-----
Flux v2.4.0 is a feature release. Users are encouraged to upgrade for
the best experience.

For a comprehensive overview of new features and API changes included
in this release, please refer to the Announcing Flux 2.4 GA blog post.

This release marks the General Availability (GA) of Flux Bucket API.
The Bucket v1 API comes with new features including: proxy support,
mTLS and custom STS configuration for AWS S3 and MinIO LDAP
authentication.

The GitRepository v1 API gains support for OIDC authentication.
Starting with this version, you can authenticate against Azure DevOps
repositories using AKS Workload Identity.

The OCIRepository v1beta2 API gains support for proxy configuration
thus allowing dedicated HTTP/S Proxy authentication on multi-tenant
Kubernetes clusters.

The HelmRelease v2 API gains support for disabling JSON schema
validation of the Helm release values during installation and upgrade.
And allows adopting existing Kubernetes resources during Helm release
installation.

The Flux controllers are now built with Go 1.23 and their dependencies
have been updated to Kubernetes 1.31, Helm 3.16, SOPS 3.9 Cosign 2.4
and Notation 1.2.

2.3.0
-----
Flux v2.3.0 is a feature release. Users are encouraged to upgrade for
the best experience.

For a comprehensive overview of new features and API changes included
in this release, please refer to the Announcing Flux 2.3 GA blog post.

This release marks the General Availability (GA) of Flux Helm features
and APIs, including helm-controller, the HelmRelease, HelmChart, and
HelmRepository APIs.

The HelmRepository v2 API comes with new features, such as the ability
to reference Helm charts from OCIRepository sources, reuse existing
HelmChart resources, and verify the integrity of Helm chart artifacts
signed with Notary Notation.

2.2.3
-----
Flux v2.2.3 is a patch release which comes with various fixes and
improvements. Users are encouraged to upgrade for the best experience.

2.2.2
-----
Flux v2.2.2 is a patch release that addresses an issue with the label
selector sharding functionality in the helm-controller. Users are
encouraged to upgrade for the best experience.

2.2.1
-----
Flux v2.2.1 is a patch release that comes with fixes to API backwards
compatibility.

2.2.0
-----
Flux v2.2.0 is a feature release. Users are encouraged to upgrade for
the best experience.

The Flux CLI and controllers have been updated to Kustomize v5.3.0 and
Kubernetes v1.28.4.

Flux helm-controller's reconciliation model underwent a significant
overhaul, addressing persistent issues such as the automatic recovery
of releases stuck in a pending state. In addition, it improves the
observability of the release status, and it introduces the ability to
enable drift detection on a per-object basis. For more details on the
helm-controller improvements, please see the Announcing Flux 2.2 GA
blog post.

The Flux CLI can now be used to force or reset the reconciliation state
of a HelmRelease v2beta2 object using flux reconcile hr --force and
flux reconcile hr --reset.

Flux CLI comes with support for bootstrapping Gitea repositories and
adds guardrails to flux install and flux bootstrap to protect users
from destructive operations. The flux version and flux check commands
now print the Flux distribution version deployed on the cluster.

The Flux alerting capabilities have been extended with NATS and
Bitbucket Server & Data Center support.

Starting with this release, Flux minor versions are benchmark to
measure the Mean Time To Production (MTTP). The results for this
version can be found at github.com/fluxcd/flux-benchmark.

2.1.2
-----
Flux v2.1.2 is a patch release which comes with various fixes. Users
are encouraged to upgrade for the best experience.