librecad: Merge recent changes in pkgsrc, and adjust C++
  - Merge changes from cad/librecad
    - boost issues (change to c++17, per upstream, and simplify)
    - don't reset MAINTAINER
  - Comment out boost-lib prefix stuff, with an explanation for
    discussion.  But it seems to build fine.

Builds ok and runs on NetBSD 10 amd64.

Note wip author in COMMIT_MSG.

A significant remaining issue is that the tarball I downloaded is
different (sha1, length) that the one in distinfo.  I am intentionally
not changing distinfo in this commit.  (Probably Atsushi and I should
compare our tarballs and ask upstream.)

Absent boost path, tarball and checking with $MAINTAINER, I think this
is ready.