rust184: tweak this so that it builds natively on NetBSD/macppc 10.0.
This is foremost to deal with an assert triggered in liblzma:

assertion "memcmp(mf_ptr(mf) - 1, mf_ptr(mf) - matches[i].dist - 2, matches[i].len) == 0" failed: file "xz-5.2/src/liblzma/lz/lz_encoder_mf.c", line 40, function "lzma_mf_find"

which of course appears at the very end of the build during the
rust-installer use.

This error is seen with both the in-tree gcc and with pkgsrc gcc12,
but forcing the use of pkgsrc gcc14 on NetBSD/powerpc fixes this problem.

That, however, gives rise to another patch being needed to the
configuration of the lzma-sys-0.1.20 crate, and a corresponding
checksum tweak, to deal with

 cargo:warning=xz-5.2/src/common/mythread.h: In function 'mythread_condtime_set':
 cargo:warning=xz-5.2/src/common/mythread.h:312:9: error: implicit declaration of function 'gettimeofday' [-Wimplicit-function-declaration]
 cargo:warning=  312 |         gettimeofday(&now, NULL);
 cargo:warning=      |         ^~~~~~~~~~~~

and we just add NetBSD to the "__sun" handling in config.h,
and tell the code that NetBSD has clock_gettime() available.
I am uncertain why this was triggered by using gcc14, and not
seen with the earlier versions.