rust184: deal with getmmntinfo() on NetBSD.
This applies https://github.com/rust-lang/libc/pull/4265/files
to deal with the versioned data structure / ABI for this libc call.

And while we're at it, carry forward the solaris patch to all
the libc crates built as part of building the rust compiler, as
well as the patch for getmntinfo().