jitsi-meet: Resolve TODO about resource limits
I built this on a machine with 12000 MB of RAM (yes, I know that's not
quite 12 GB), including WRKDIR on tmpfs, without issues and have noted
that in the Makefile.  But, with 6000 MB there was unacceptable
paging.  This package is therefore not unusual compared to many others
that won't build without lots of resources.

Adjust TODO:
  - note need to remediate google analytics
  - note that chat/jitsi-meet/Makefile.common should probably be split