jitsi-meet: Adjust TODO
On investigating adding branding.json as a config file and not finding
it in PLIST or the sources, drop that item.

After noticing google css implying that a google logo is loaded,
suggesting there are third-party page loads for external auth, enhance
the entry for removing phone-home of GA to include all default-on
third-party page loads.