kicad-*: Update to 9.0.0
KiCad 9.0.0 (2025-02-20)

Most, if not all, of the new features as well as usability
improvements in KiCad 9 are described in a running thread on the KiCad
user forum. This thread is being kept reasonably up to date so users
running nightly builds know about new features to take for a test spin
and provide bug reports and feedback. Wayne Stambaugh also presented
them in his FOSDEM 2025 talk which is available to watch on the FOSDEM
website. This post highlights some of the major changes, but is not a
complete list. There were also hundreds of bug fixes, performance
improvements, and other smaller changes since KiCad 8. A full list of
issues addressed in KiCad 9 can be found on the milestone page on
GitLab.